Refactor: Move SwearJarListApiResponse, RecentSwearsApiResponse and SwearJarTrendApiResponse to apiTypes
