+++ frontend/app/swearjar/[id]/view/page.tsx
@@ -4,33 +4,12 @@ import {
   QueryClient,
 } from '@tanstack/react-query'
 import { fetcher } from "@/lib/utils";
-import { SwearJarBase, Swear, User } from '@/lib/types';
+import { SwearJarListApiResponse, RecentSwearsApiResponse, SwearJarTrendApiResponse } from '@/lib/apiTypes';
 
 
 import DefaultContentLayout from "@/components/layout/content";
 import MainContent from "./mainContent";
 import ErrorAlert from '@/components/shared/ErrorAlert';
-
-export interface SwearJarApiResponse {
-  msg: string;
-  swearJar: SwearJarBase;
-}
-
-export interface RecentSwearsApiResponse {
-  msg: string;
-  data: {
-    swears: Swear[];
-    users: { [key: string]: User };
-  }
-}
-
-export interface SwearJarTrendApiResponse {
-  msg: string;
-  data: {
-    label: string;
-    [key: string]: number | string; // This allows for one or more users with their respective swear counts
-  }[];
-}
 
 export default async function SwearJarPage({ params }: { params: { id: string } }) {
   const queryClient = new QueryClient()
@@ -38,9 +17,9 @@ export default async function SwearJarPage({ params }: { params: { id: string } 
   try {
     // Fetch all data in parallel
     await Promise.all([
-      queryClient.prefetchQuery<SwearJarApiResponse>({
+      queryClient.prefetchQuery<SwearJarListApiResponse>({
         queryKey: [`swearjar?id=${params.id}`],
-        queryFn: () => fetcher<SwearJarApiResponse>(`/api/swearjar?id=${params.id}`),
+        queryFn: () => fetcher<SwearJarListApiResponse>(`/api/swearjar?id=${params.id}`),
       }),
       queryClient.prefetchQuery<RecentSwearsApiResponse>({
         queryKey: [`swear?id=${params.id}`],
