+++ frontend/app/swearjar/create/mainContent.tsx
@@ -1,9 +1,9 @@
-import CreateSwearJarForm from "@/components/app/swearjar/forms/SwearJarForm";
+import SwearJarForm from "@/components/app/swearjar/forms/SwearJarForm";
 
 export default function MainContent() {
     return (
         <main>
-            <CreateSwearJarForm />
+            <SwearJarForm />
         </main>
     )
 }
